requirements.txt
vk==3.0
vk-api==11.9.9
python-dotenv~=0.21.0